+++ resources/js/layouts/LoginTemplate.vue
@@ -1,0 +1,71 @@
+<template>
+  <span>
+    <header>
+      <navbar :color="color" logo="brand-logo" :title="title" />
+    </header>
+    <main>
+      <div class="container">
+        <div class="row">
+          <slot></slot>
+        </div>
+      </div>
+    </main>
+    <footerbar :color="color" :title="title">
+      <li><a class="grey-text text-lighten-3" href="#">Home</a></li>
+    </footerbar>
+  </span>
+</template>
+
+<script>
+import Navbar from "@/components/layouts/Navbar";
+import Footerbar from "@/components/layouts/Footerbar";
+import GridVue from "@/components/layouts/GridVue";
+import CardPhotoVue from "@/components/user/CardPhotoVue";
+import CardMenuItensVue from "@/components/user/CardMenuItensVue";
+
+export default {
+  name: 'LoginTemplate',
+
+
+  components: {
+    CardMenuItensVue,
+    CardPhotoVue,
+    GridVue,
+    Navbar,
+    Footerbar
+  },
+
+  data() {
+    return {
+      title: 'Lorem ipsum',
+      color: 'blue-grey darken-2',
+    }
+  },
+
+  created() {
+    let userLogged = sessionStorage.getItem('user');
+
+    if (userLogged) {
+      this.user = JSON.parse(userLogged);
+      this.$router.push({name:'Home'})
+    }
+  },
+}
+</script>
+
+<style>
+header {
+  z-index: 10;
+  position: fixed;
+  width: 100%;
+}
+
+main {
+  margin-top: 6rem;
+}
+
+#user-row {
+  width: clamp(200px, 20%, 300px);
+  position: fixed;
+}
+</style>