+++ CHANGELOG.md
@@ -10,3 +10,16 @@
   - Poster
 - CardPhotoVue (user's photo)
 - CardMenuItensVue (user's items)
+
+## [1.1.0] - 2021-01-13
+### Added
+- LoginTemplate (layout)
+- SiteTemplate (layout)
+- Login (page)
+  - LoginForm (form)
+  - RegisterForm (form)
+- Laravel (PHP framework)
+- "Backend to login works"
+- "Session storage user"
+- "User's authentication"
+